Feature/react integration

+++ ResourseAPI/Controllers/ResourcesController.cs
@@ -1,6 +1,8 @@
 ﻿using System.Collections.Generic;
 using ResourseAPI.Models;
 using ResourseAPI.Services;
+using AuthAPI.Models;
+using AuthAPI.Services;
 using Microsoft.AspNetCore.Mvc;
 using ResourseAPI.Utility;
 
@@ -13,31 +15,74 @@ namespace ResourseAPI.Controllers
     {
         private readonly ResourseService _resourseService;
         private readonly SubscribeService _subscribeService;
-        private readonly UserService _userService;
+        private readonly UsersService _usersService;
+        private readonly SessionsService _sessionsService;
 
-        public ResourceController(ResourseService resourseService, SubscribeService subscribeService, UserService userService)
+        public ResourceController(ResourseService resourseService, SubscribeService subscribeService, UsersService usersService, SessionsService sessionsService)
         {
             _resourseService = resourseService;
             _subscribeService = subscribeService;
-            _userService = userService;
+            _usersService = usersService;
+            _sessionsService = sessionsService;
+        }
+
+        private bool isAuthorized(string Authorization)
+        {
+            return this._sessionsService.GetSession(Authorization) != null;
+        }
+        
+        private bool isUserATeacher(string Authorization)
+        {
+            var user = this._sessionsService.GetUser(Authorization);
+
+            return user != null
+                ? user.Role == "teacher"
+                : false;
         }
 
         [HttpGet(Name = "Get")]
-        public ActionResult<List<Resourse>> Get(string id, string courseId)
+        public ActionResult<List<Resourse>> Get([FromHeaderAttribute] string Authorization, string id, string courseId)
         {
-            if ((id != null && !id.Equals("")) || (courseId != null && !courseId.Equals("")))
+            if (isAuthorized(Authorization))
             {
-                List<Resourse> resourse = null;
-                if (id != null && !id.Equals(""))
+                if ((id != null && !id.Equals("")) || (courseId != null && !courseId.Equals("")))
                 {
-                    resourse = new List<Resourse>();
-                    Resourse resource = _resourseService.Get(id);
-                    if (resource != null) resourse.Add(resource);
+                    List<Resourse> resourse = null;
+                    if (id != null && !id.Equals(""))
+                    {
+                        resourse = new List<Resourse>();
+                        Resourse resource = _resourseService.Get(id);
+                        if (resource != null) resourse.Add(resource);
+                    }
+                    else
+                    {
+                        resourse = _resourseService.GetByCourseId(courseId);
+                    }
+
+                    if (resourse == null)
+                    {
+                        return NotFound();
+                    }
+
+                    return resourse;
                 }
                 else
                 {
-                    resourse = _resourseService.GetByCourseId(courseId);
+                    return _resourseService.Get();
                 }
+            }
+            else
+            {
+                return Unauthorized();
+            }
+        }
+
+        [HttpGet("{id:length(24)}", Name = "GetResourse")]
+        public ActionResult<Resourse> GetById([FromHeaderAttribute] string Authorization, string id)
+        {
+            if (isAuthorized(Authorization))
+            {
+                var resourse = _resourseService.Get(id);
 
                 if (resourse == null)
                 {
@@ -48,60 +93,67 @@ namespace ResourseAPI.Controllers
             }
             else
             {
-                return _resourseService.Get();
+                return Unauthorized();
             }
-        }
-
-        [HttpGet("{id:length(24)}", Name = "GetResourse")]
-        public ActionResult<Resourse> GetById(string id)
-        {
-            var resourse = _resourseService.Get(id);
-
-            if (resourse == null)
-            {
-                return NotFound();
-            }
-
-            return resourse;
         }
 
         // localhost:5005/api/resourses/create
         [HttpPost("create")]
-        public ActionResult<Resourse> Create([FromBody] Resourse resourse)
+        public ActionResult<Resourse> Create([FromHeaderAttribute] string Authorization, [FromBody] Resourse resourse)
         {
-            User user;
-            List<Subscribe> subscribes;
-            string message;
-
-           subscribes =  _subscribeService.GetByCourseId(resourse.CourseId);
-
-            if (subscribes == null)
+            if (isAuthorized(Authorization) && isUserATeacher(Authorization))
             {
-                return NotFound();
-            }
-            foreach (var subscribe in subscribes)
-            {
-                user = _userService.Get(subscribe.ID_User);
-     
-                if (user == null)
+                User user;
+                List<Subscribe> subscribes;
+                string message;
+
+                subscribes = _subscribeService.GetByCourseId(resourse.CourseId);
+
+                if (subscribes == null)
                 {
                     return NotFound();
                 }
 
-                message = "S-a incarcat o noua resursa cu titlul " + resourse.Title + "\n"
-                          + "Cu descrierea: " + resourse.Description + "\n"
-                          + "Resursa poate fi accesata la: " + resourse.Link;
-        
-                MailHelper.SendMail(user.Email, message);
+                foreach (var subscribe in subscribes)
+                {
+                    user = _usersService.Get(subscribe.ID_User);
 
+                    if (user == null)
+                    {
+                        return NotFound();
+                    }
+
+
+                    message = "S-a incarcat o noua resursa cu titlul " + resourse.Title + "\n"
+                            + "Cu descrierea: " + resourse.Description + "\n"
+                            + "Resursa poate fi accesata la: " + resourse.Link;
+
+
+
+                    MailHelper.SendMail(user.Email, message);
+
+                }
+
+                _resourseService.Create(resourse);
+
+                return CreatedAtRoute("GetResourse", new { id = resourse._id.ToString() }, resourse);
             }
-            _resourseService.Create(resourse);
-            return CreatedAtRoute("Get", new { id = resourse._id.ToString() }, resourse);
+            else
+            {
+                return Unauthorized();
+            }
         }
 
         [HttpPut("{id:length(24)}")]
-        public IActionResult Update(string id, Resourse resourseIn)
+        public IActionResult Update([FromHeaderAttribute] string Authorization, string id, Resourse resourseIn)
         {
+            if (isAuthorized(Authorization))
+            {
+            }
+            else
+            {
+                return Unauthorized();
+            }
             var resourse = _resourseService.Get(id);
 
             if (resourse == null)
@@ -115,24 +167,30 @@ namespace ResourseAPI.Controllers
         }
 
         [HttpDelete("{id:length(24)}")]
-        public IActionResult Delete(string id)
+        public IActionResult Delete([FromHeaderAttribute] string Authorization, string id)
         {
-            var resourse = _resourseService.Get(id);
-
-            if (resourse == null)
+            if (isAuthorized(Authorization))
             {
-                return NotFound();
+                var resourse = _resourseService.Get(id);
+
+                if (resourse == null)
+                {
+                    return NotFound();
+                }
+
+                _resourseService.Remove(resourse._id);
+
+                return NoContent();
             }
-
-            _resourseService.Remove(resourse._id);
-
-            return NoContent();
+            else
+            {
+                return Unauthorized();
+            }
         }
 
         // [HttpPost("upload")]
         // public async Task<IActionResult> Post(List<Microsoft.AspNetCore.Http.IFormFile> files)
         // {
-        //     System.Console.WriteLine("******************");
         //     long size = files.Sum(f => f.Length);
 
         //     // full path to file in temp location
